Tidy up and simplify.
